--- docs/source/api/sample/Dixon/_r11.rst
@@ -1,6 +1,0 @@
-_r11
-====
-
-.. currentmodule:: pycafee.sample.dixon.Dixon
-
-.. autofunction:: pycafee.sample.dixon.Dixon._r11
