--- docs/source/api/sample/Dixon/_r12.rst
@@ -1,6 +1,0 @@
-_r12
-====
-
-.. currentmodule:: pycafee.sample.dixon.Dixon
-
-.. autofunction:: pycafee.sample.dixon.Dixon._r12
