--- docs/source/api/sample/Dixon/_r20.rst
@@ -1,6 +1,0 @@
-_r20
-====
-
-.. currentmodule:: pycafee.sample.dixon.Dixon
-
-.. autofunction:: pycafee.sample.dixon.Dixon._r20
