--- docs/source/api/sample/Dixon/_r21.rst
@@ -1,6 +1,0 @@
-_r21
-====
-
-.. currentmodule:: pycafee.sample.dixon.Dixon
-
-.. autofunction:: pycafee.sample.dixon.Dixon._r21
