--- docs/source/api/sample/Dixon/_r22.rst
@@ -1,6 +1,0 @@
-_r22
-====
-
-.. currentmodule:: pycafee.sample.dixon.Dixon
-
-.. autofunction:: pycafee.sample.dixon.Dixon._r22
