--- docs/source/api/sample/Dixon/fit.rst
@@ -1,6 +1,0 @@
-fit
-===
-
-.. currentmodule:: pycafee.sample.dixon.Dixon
-
-.. autofunction:: pycafee.sample.dixon.Dixon.fit
