--- docs/source/api/sample/Dixon/get_critical_value.rst
@@ -1,6 +1,0 @@
-get_critical_value
-==================
-
-.. currentmodule:: pycafee.sample.dixon.Dixon
-
-.. autofunction:: pycafee.sample.dixon.Dixon.get_critical_value
