+++ docs/source/api/sample/Outliers/Dixon/_r10.rst
@@ -1,0 +1,6 @@
+_r10
+====
+
+.. currentmodule:: pycafee.sample.outliers.Dixon
+
+.. autofunction:: pycafee.sample.outliers.Dixon._r10
