+++ docs/source/api/sample/Outliers/Dixon/_r11.rst
@@ -1,0 +1,6 @@
+_r11
+====
+
+.. currentmodule:: pycafee.sample.outliers.Dixon
+
+.. autofunction:: pycafee.sample.outliers.Dixon._r11
