+++ docs/source/api/sample/Outliers/Dixon/_r20.rst
@@ -1,0 +1,6 @@
+_r20
+====
+
+.. currentmodule:: pycafee.sample.outliers.Dixon
+
+.. autofunction:: pycafee.sample.outliers.Dixon._r20
