+++ docs/source/api/sample/Outliers/Dixon/_r21.rst
@@ -1,0 +1,6 @@
+_r21
+====
+
+.. currentmodule:: pycafee.sample.outliers.Dixon
+
+.. autofunction:: pycafee.sample.outliers.Dixon._r21
