+++ docs/source/api/sample/Outliers/Dixon/fit.rst
@@ -1,0 +1,6 @@
+fit
+===
+
+.. currentmodule:: pycafee.sample.outliers.Dixon
+
+.. autofunction:: pycafee.sample.outliers.Dixon.fit
