+++ docs/source/api/sample/Outliers/Dixon/get_critical_value.rst
@@ -1,0 +1,6 @@
+get_critical_value
+==================
+
+.. currentmodule:: pycafee.sample.outliers.Dixon
+
+.. autofunction:: pycafee.sample.outliers.Dixon.get_critical_value
