+++ docs/source/api/sample/Outliers/Dixon/index.rst
