+++ docs/source/api/sample/Outliers/index.rst
@@ -8,6 +8,7 @@ Outliers
 
    ZScore/index
    ModifiedZScore/index
+   Dixon/index
 
 |
 |
