+++ docs/source/api/sample/index.rst
@@ -8,7 +8,6 @@ Sample
 
    Sample/index
    StudentDistribution/index
-   Dixon/index
    Outliers/index
 
 |
