+++ tests/sample/outliers/Dixon/test_fit.py
@@ -6,16 +6,16 @@
 
 --------------------------------------------------------------------------------
 Command to run at the prompt:
-    python -m unittest -v tests/sample/Dixon/test_fit.py
+    python -m unittest -v tests/sample/outliers/Dixon/test_fit.py
     or
-    python -m unittest -b tests/sample/Dixon/test_fit.py
+    python -m unittest -b tests/sample/outliers/Dixon/test_fit.py
 
 --------------------------------------------------------------------------------
 """
 
 import os
 import unittest
-from pycafee.sample.dixon import Dixon
+from pycafee.sample.outliers import Dixon
 import numpy as np
 os.system('cls')
